+++ app/commons/log_preparation.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from app.utils import utils
+from app.utils import utils, text_processing
 from datetime import datetime
 from app.commons.log_merger import LogMerger
 
@@ -23,13 +23,13 @@ class LogPreparation:
         self.log_merger = LogMerger()
 
     def clean_message(self, message):
-        message = utils.replace_tabs_for_newlines(message)
-        message = utils.fix_big_encoded_urls(message)
-        message = utils.remove_generated_parts(message)
-        message = utils.remove_guid_uids_from_text(message)
-        message = utils.clean_html(message)
-        message = utils.delete_empty_lines(message)
-        message = utils.leave_only_unique_lines(message)
+        message = text_processing.replace_tabs_for_newlines(message)
+        message = text_processing.fix_big_encoded_urls(message)
+        message = text_processing.remove_generated_parts(message)
+        message = text_processing.remove_guid_uids_from_text(message)
+        message = text_processing.clean_html(message)
+        message = text_processing.delete_empty_lines(message)
+        message = text_processing.leave_only_unique_lines(message)
         return message
 
     def _create_log_template(self):
@@ -79,7 +79,7 @@ class LogPreparation:
         log_template["_source"]["unique_id"] = test_item.uniqueId
         log_template["_source"]["test_case_hash"] = test_item.testCaseHash
         log_template["_source"]["is_auto_analyzed"] = test_item.isAutoAnalyzed
-        log_template["_source"]["test_item_name"] = utils.preprocess_test_item_name(test_item.testItemName)
+        log_template["_source"]["test_item_name"] = text_processing.preprocess_test_item_name(test_item.testItemName)
         log_template["_source"]["issue_type"] = self.transform_issue_type_into_lowercase(
             test_item.issueType)
         log_template["_source"]["start_time"] = datetime(
@@ -89,46 +89,47 @@ class LogPreparation:
     def _fill_log_fields(self, log_template, log, number_of_lines):
         cleaned_message = self.clean_message(log.message)
 
-        test_and_methods = utils.find_test_methods_in_text(cleaned_message)
-        message = utils.first_lines(cleaned_message, number_of_lines)
-        message = utils.replace_text_pieces(message, test_and_methods)
+        test_and_methods = text_processing.find_test_methods_in_text(cleaned_message)
+        message = text_processing.first_lines(cleaned_message, number_of_lines)
+        message = text_processing.replace_text_pieces(message, test_and_methods)
         message_without_params = message
-        message = utils.delete_empty_lines(utils.sanitize_text(message))
+        message = text_processing.delete_empty_lines(text_processing.sanitize_text(message))
 
-        message_without_params = utils.clean_from_urls(message_without_params)
-        message_without_params = utils.clean_from_paths(message_without_params)
-        message_without_params = utils.clean_from_params(message_without_params)
-        message_without_params = utils.remove_starting_datetime(message_without_params)
-        message_without_params = utils.sanitize_text(message_without_params)
-        message_without_params_and_brackets = utils.clean_from_brackets(message_without_params)
+        message_without_params = text_processing.clean_from_urls(message_without_params)
+        message_without_params = text_processing.clean_from_paths(message_without_params)
+        message_without_params = text_processing.clean_from_params(message_without_params)
+        message_without_params = text_processing.remove_starting_datetime(message_without_params)
+        message_without_params = text_processing.sanitize_text(message_without_params)
+        message_without_params_and_brackets = text_processing.clean_from_brackets(message_without_params)
 
-        detected_message, stacktrace = utils.detect_log_description_and_stacktrace(cleaned_message)
+        detected_message, stacktrace = text_processing.detect_log_description_and_stacktrace(cleaned_message)
 
         detected_message_without_params = detected_message
-        urls = " ".join(utils.extract_urls(detected_message_without_params))
-        detected_message_without_params = utils.clean_from_urls(detected_message_without_params)
-        paths = " ".join(utils.extract_paths(detected_message_without_params))
-        detected_message_without_params = utils.clean_from_paths(detected_message_without_params)
-        potential_status_codes = " ".join(utils.get_potential_status_codes(detected_message_without_params))
-        detected_message_without_params = utils.replace_text_pieces(
+        urls = " ".join(text_processing.extract_urls(detected_message_without_params))
+        detected_message_without_params = text_processing.clean_from_urls(detected_message_without_params)
+        paths = " ".join(text_processing.extract_paths(detected_message_without_params))
+        detected_message_without_params = text_processing.clean_from_paths(detected_message_without_params)
+        potential_status_codes = " ".join(
+            text_processing.get_potential_status_codes(detected_message_without_params))
+        detected_message_without_params = text_processing.replace_text_pieces(
             detected_message_without_params, test_and_methods)
-        detected_message = utils.replace_text_pieces(detected_message, test_and_methods)
-        detected_message_without_params = utils.remove_starting_datetime(detected_message_without_params)
+        detected_message = text_processing.replace_text_pieces(detected_message, test_and_methods)
+        detected_message_without_params = text_processing.remove_starting_datetime(detected_message_without_params)
         detected_message_wo_urls_and_paths = detected_message_without_params
 
-        message_params = " ".join(utils.extract_message_params(detected_message_without_params))
-        detected_message_without_params = utils.clean_from_params(detected_message_without_params)
-        detected_message_without_params = utils.sanitize_text(detected_message_without_params)
-        detected_message_without_params_and_brackets = utils.clean_from_brackets(
+        message_params = " ".join(text_processing.extract_message_params(detected_message_without_params))
+        detected_message_without_params = text_processing.clean_from_params(detected_message_without_params)
+        detected_message_without_params = text_processing.sanitize_text(detected_message_without_params)
+        detected_message_without_params_and_brackets = text_processing.clean_from_brackets(
             detected_message_without_params)
 
-        detected_message_with_numbers = utils.remove_starting_datetime(detected_message)
-        detected_message_only_numbers = utils.find_only_numbers(detected_message_with_numbers)
-        detected_message = utils.sanitize_text(detected_message)
-        stacktrace = utils.sanitize_text(stacktrace)
-        found_exceptions = utils.get_found_exceptions(detected_message)
-        found_exceptions_extended = utils.enrich_found_exceptions(found_exceptions)
-        found_test_methods = utils.enrich_text_with_method_and_classes(" ".join(test_and_methods))
+        detected_message_with_numbers = text_processing.remove_starting_datetime(detected_message)
+        detected_message_only_numbers = text_processing.find_only_numbers(detected_message_with_numbers)
+        detected_message = text_processing.sanitize_text(detected_message)
+        stacktrace = text_processing.sanitize_text(stacktrace)
+        found_exceptions = text_processing.get_found_exceptions(detected_message)
+        found_exceptions_extended = text_processing.enrich_found_exceptions(found_exceptions)
+        found_test_methods = text_processing.enrich_text_with_method_and_classes(" ".join(test_and_methods))
 
         log_template["_id"] = log.logId
         log_template["_source"]["log_time"] = datetime(*log.logTime[:6]).strftime("%Y-%m-%d %H:%M:%S")
@@ -136,9 +137,9 @@ class LogPreparation:
         log_template["_source"]["cluster_message"] = log.clusterMessage
         log_template["_source"]["cluster_with_numbers"] = utils.extract_clustering_setting(log.clusterId)
         log_template["_source"]["log_level"] = log.logLevel
-        log_template["_source"]["original_message_lines"] = utils.calculate_line_number(cleaned_message)
+        log_template["_source"]["original_message_lines"] = text_processing.calculate_line_number(cleaned_message)
         log_template["_source"]["original_message_words_number"] = len(
-            utils.split_words(cleaned_message, split_urls=False))
+            text_processing.split_words(cleaned_message, split_urls=False))
         log_template["_source"]["message"] = message
         log_template["_source"]["detected_message"] = detected_message
         log_template["_source"]["detected_message_with_numbers"] = detected_message_with_numbers
@@ -150,15 +151,15 @@ class LogPreparation:
         log_template["_source"]["found_exceptions"] = found_exceptions
         log_template["_source"]["found_exceptions_extended"] = found_exceptions_extended
         log_template["_source"]["detected_message_extended"] =\
-            utils.enrich_text_with_method_and_classes(detected_message)
+            text_processing.enrich_text_with_method_and_classes(detected_message)
         log_template["_source"]["detected_message_without_params_extended"] =\
-            utils.enrich_text_with_method_and_classes(detected_message_without_params)
+            text_processing.enrich_text_with_method_and_classes(detected_message_without_params)
         log_template["_source"]["stacktrace_extended"] =\
-            utils.enrich_text_with_method_and_classes(stacktrace)
+            text_processing.enrich_text_with_method_and_classes(stacktrace)
         log_template["_source"]["message_extended"] =\
-            utils.enrich_text_with_method_and_classes(message)
+            text_processing.enrich_text_with_method_and_classes(message)
         log_template["_source"]["message_without_params_extended"] =\
-            utils.enrich_text_with_method_and_classes(message_without_params)
+            text_processing.enrich_text_with_method_and_classes(message_without_params)
         log_template["_source"]["whole_message"] = detected_message_wo_urls_and_paths + " \n " + stacktrace
         log_template["_source"]["detected_message_without_params_and_brackets"] =\
             detected_message_without_params_and_brackets
@@ -174,8 +175,8 @@ class LogPreparation:
                       "stacktrace_extended", "message_extended", "message_without_params_extended",
                       "detected_message_without_params_and_brackets",
                       "message_without_params_and_brackets"]:
-            log_template["_source"][field] = utils.leave_only_unique_lines(log_template["_source"][field])
-            log_template["_source"][field] = utils.clean_colon_stacking(log_template["_source"][field])
+            log_template["_source"][field] = text_processing.leave_only_unique_lines(log_template["_source"][field])
+            log_template["_source"][field] = text_processing.clean_colon_stacking(log_template["_source"][field])
         return log_template
 
     def _prepare_log(self, launch, test_item, log, project):
@@ -191,7 +192,7 @@ class LogPreparation:
         log_template["_source"]["test_item"] = test_item_info.testItemId
         log_template["_source"]["unique_id"] = test_item_info.uniqueId
         log_template["_source"]["test_case_hash"] = test_item_info.testCaseHash
-        log_template["_source"]["test_item_name"] = utils.preprocess_test_item_name(
+        log_template["_source"]["test_item_name"] = text_processing.preprocess_test_item_name(
             test_item_info.testItemName)
         log_template["_source"]["is_auto_analyzed"] = False
         log_template["_source"]["issue_type"] = ""
@@ -216,9 +217,9 @@ class LogPreparation:
                     if log.logLevel < utils.ERROR_LOGGING_LEVEL or not log.message.strip():
                         continue
                     clean_message = self.clean_message(log.message)
-                    det_message, stacktrace = utils.detect_log_description_and_stacktrace(
+                    det_message, stacktrace = text_processing.detect_log_description_and_stacktrace(
                         clean_message)
-                    for word in utils.split_words(stacktrace):
+                    for word in text_processing.split_words(stacktrace):
                         if "." in word and len(word.split(".")) > 2:
                             log_words[word] = 1
         return log_words, project
@@ -227,34 +228,34 @@ class LogPreparation:
         log_template = self._create_log_template()
         log_template = self._fill_launch_test_item_fields(log_template, launch, test_item, project)
         cleaned_message = self.clean_message(log.message)
-        detected_message, stacktrace = utils.detect_log_description_and_stacktrace(
+        detected_message, stacktrace = text_processing.detect_log_description_and_stacktrace(
             cleaned_message)
-        test_and_methods = utils.find_test_methods_in_text(cleaned_message)
-        detected_message = utils.replace_text_pieces(detected_message, test_and_methods)
-        stacktrace = utils.sanitize_text(stacktrace)
-        message = utils.first_lines(cleaned_message, -1)
-        message = utils.sanitize_text(message)
+        test_and_methods = text_processing.find_test_methods_in_text(cleaned_message)
+        detected_message = text_processing.replace_text_pieces(detected_message, test_and_methods)
+        stacktrace = text_processing.sanitize_text(stacktrace)
+        message = text_processing.first_lines(cleaned_message, -1)
+        message = text_processing.sanitize_text(message)
         log_template["_id"] = log.logId
         log_template["_source"]["cluster_id"] = str(log.clusterId)
         log_template["_source"]["cluster_message"] = log.clusterMessage
         log_template["_source"]["log_level"] = log.logLevel
-        log_template["_source"]["original_message_lines"] = utils.calculate_line_number(
+        log_template["_source"]["original_message_lines"] = text_processing.calculate_line_number(
             cleaned_message)
         log_template["_source"]["original_message_words_number"] = len(
-            utils.split_words(cleaned_message, split_urls=False))
-        detected_message_wo_urls_and_paths = utils.clean_from_urls(detected_message)
-        detected_message_wo_urls_and_paths = utils.clean_from_paths(detected_message_wo_urls_and_paths)
-        detected_message_wo_urls_and_paths = utils.remove_starting_datetime(
+            text_processing.split_words(cleaned_message, split_urls=False))
+        detected_message_wo_urls_and_paths = text_processing.clean_from_urls(detected_message)
+        detected_message_wo_urls_and_paths = text_processing.clean_from_paths(detected_message_wo_urls_and_paths)
+        detected_message_wo_urls_and_paths = text_processing.remove_starting_datetime(
             detected_message_wo_urls_and_paths)
         log_template["_source"]["message"] = message
         log_template["_source"]["detected_message"] = detected_message_wo_urls_and_paths
         log_template["_source"]["detected_message_with_numbers"] = detected_message_wo_urls_and_paths
         log_template["_source"]["stacktrace"] = stacktrace
         potential_status_codes = " ".join(
-            utils.get_potential_status_codes(detected_message_wo_urls_and_paths))
+            text_processing.get_potential_status_codes(detected_message_wo_urls_and_paths))
         log_template["_source"]["potential_status_codes"] = potential_status_codes
-        log_template["_source"]["found_exceptions"] = utils.get_found_exceptions(detected_message)
-        log_template["_source"]["whole_message"] = utils.delete_empty_lines(
+        log_template["_source"]["found_exceptions"] = text_processing.get_found_exceptions(detected_message)
+        log_template["_source"]["whole_message"] = text_processing.delete_empty_lines(
             detected_message_wo_urls_and_paths + " \n " + stacktrace)
         return log_template
 
@@ -278,7 +279,7 @@ class LogPreparation:
                 number_of_log_lines = number_of_lines
                 if log["_source"]["is_merged"]:
                     number_of_log_lines = -1
-                log_message = utils.prepare_message_for_clustering(
+                log_message = text_processing.prepare_message_for_clustering(
                     log["_source"]["whole_message"], number_of_log_lines, clean_numbers)
                 if not log_message.strip():
                     continue
