+++ requirements-dev.txt
@@ -1,3 +1,5 @@
+pika-stubs==0.1.3
+
 pytest==7.4.0
 pytest-cov==4.1.0
 flake8==5.0.4
